Move LocalEntity to data package
